+++ tests/test_context.py
@@ -7,7 +7,9 @@ import pytest
 from dipdup.config import DipDupConfig
 from dipdup.dipdup import DipDup
 from dipdup.enums import ReindexingReason
+from dipdup.exceptions import ContractAlreadyExistsError
 from dipdup.exceptions import ReindexingRequiredError
+from dipdup.models import Contract
 from dipdup.models import Schema
 
 
@@ -39,3 +41,24 @@ async def test_reindex_field(dummy_dipdup: DipDup) -> None:
     # Assert
     schema = await Schema.filter().get()
     assert schema.reindex == ReindexingReason.manual
+
+
+async def test_add_contract(dummy_dipdup: DipDup) -> None:
+    ctx = dummy_dipdup._ctx
+    await ctx.add_contract('address', 'KT1K4EwTpbvYN9agJdjpyJm4ZZdhpUNKB3F6')
+    await ctx.add_contract('code_hash', None, 'KT1K4EwTpbvYN9agJdjpyJm4ZZdhpUNKB3F6')
+
+    with pytest.raises(ContractAlreadyExistsError):
+        await ctx.add_contract('address_dup', 'KT1K4EwTpbvYN9agJdjpyJm4ZZdhpUNKB3F6')
+    with pytest.raises(ContractAlreadyExistsError):
+        await ctx.add_contract('address', 'KT1K4EwTpbvYN9agJdjpyJm4ZZdhpUNK0000')
+    with pytest.raises(ContractAlreadyExistsError):
+        await ctx.add_contract('code_hash_dup', None, 'KT1K4EwTpbvYN9agJdjpyJm4ZZdhpUNKB3F6')
+    with pytest.raises(ContractAlreadyExistsError):
+        await ctx.add_contract('code_hash', None, 'KT1K4EwTpbvYN9agJdjpyJm4ZZdhpUNK0000')
+
+    assert ctx.config.get_contract('address').address == 'KT1K4EwTpbvYN9agJdjpyJm4ZZdhpUNKB3F6'
+    assert ctx.config.get_contract('code_hash').code_hash == 'KT1K4EwTpbvYN9agJdjpyJm4ZZdhpUNKB3F6'
+
+    assert (await Contract.get(name='address')).address == 'KT1K4EwTpbvYN9agJdjpyJm4ZZdhpUNKB3F6'
+    assert (await Contract.get(name='code_hash')).address == ':KT1K4EwTpbvYN9agJdjpyJm4ZZdhpUNKB3F6'