+++ tests/test_datasources/test_tzkt.py
@@ -5,10 +5,12 @@ from typing import TypeVar
 from unittest.mock import AsyncMock
 
 import orjson as json
+import pysignalr.exceptions
 import pytest
 
 from dipdup.datasources.tzkt.models import HeadSubscription
 from dipdup.enums import MessageType
+from dipdup.exceptions import DatasourceError
 from dipdup.exceptions import FrameworkException
 from dipdup.exceptions import InvalidRequestError
 from dipdup.models import OperationData
@@ -202,3 +204,16 @@ async def test_no_content() -> None:
     async with tzkt_replay('https://api.jakartanet.tzkt.io', batch_size=1) as tzkt:
         with pytest.raises(InvalidRequestError):
             await tzkt.get_jsonschemas('KT1EHdK9asB6BtPLvt1ipKRuxsrKoQhDoKgs')
+
+
+async def test_signalr_client() -> None:
+    fail_mock = AsyncMock(side_effect=pysignalr.exceptions.ConnectionError(418))
+
+    async with tzkt_replay(batch_size=1) as tzkt:
+        tzkt._http_config.retry_sleep = 0
+        signalr_client = tzkt._get_signalr_client()
+        signalr_client.run = fail_mock  # type: ignore[assignment]
+
+        with pytest.raises(DatasourceError):
+            await tzkt.run()
+        assert fail_mock.call_count == tzkt._http_config.retry_count